# docs/swagger.yaml
basePath: /api/v1
...
host: localhost:8082
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Quantitative Contract Automated Trading System API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: QCAT API
  version: "1.0"
paths:
...
  /cache/config:
    get:
      consumes:
      - application/json
      description: Get current cache configuration and settings
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              config:
                type: object
              timestamp:
                type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Get cache configuration
      tags:
      - Cache
  /cache/counters/reset:
    post:
      consumes:
      - application/json
      description: Reset all cache monitoring counters and statistics
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
              reset_time:
                type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Reset cache counters
      tags:
      - Cache
  /cache/events:
    get:
      consumes:
      - application/json
      description: Get recent cache events and activity logs
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              count:
                type: integer
              events:
                items:
                  type: object
                type: array
              timestamp:
                type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Get cache events
      tags:
      - Cache
  /cache/fallback/force:
    post:
      consumes:
      - application/json
      description: Force the cache system to enter fallback mode
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              fallback_mode:
                type: boolean
              message:
                type: string
              reason:
                type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Force cache fallback
      tags:
      - Cache
...
  /cache/status:
    get:
      consumes:
      - application/json
      description: Get current cache status and basic statistics
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              stats:
                type: object
              status:
                type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Get cache status
      tags:
      - Cache
  /cache/test:
    post:
      consumes:
      - application/json
      description: Test cache operations including read, write, and delete
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              status:
                type: string
              test_results:
                type: object
            type: object
        "503":
          description: Service Unavailable
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Test cache operations
      tags:
      - Cache
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"